assignment1/pyproject.toml
[tool.poetry]
name = "assignment1"
version = "0.1.0"
description = "trossen pincherx 100 and kuka kinematics simulation"
authors = ["frankcholula <lu.phrank@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
roboticstoolbox-python = {extras = ["collisions"], version = "^1.1.1"}
numpy = "1.26.4"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
